feat(backoffice): staged chat attachment opens full-size before sending (#258)
The composer's pending-photo thumbnail is now a button that opens the same
ChatImageLightbox the sent bubbles use (extracted to a shared component),
so the coach can check the image before hitting Enviar.
