+++ backoffice/messages/es.json
@@ -1231,6 +1231,7 @@
       "photoCta": "Adjuntar foto",
       "photoLabel": "Foto",
       "removeAttachment": "Quitar foto",
+      "previewAttachment": "Ver foto en grande",
       "voiceCta": "Adjuntar audio",
       "voiceLabel": "Audio",
       "voiceRecordCta": "Grabar audio",
